requirements.txt
web3>=6.0.0
pyyaml>=6.0
rich>=13.0.0
click>=8.0.0
pytest>=7.0.0
